it is pretty messy, but the LCM solution was the correct one
